pyproject.toml
[project]
name = "multi-agent-support-system" 
version = "0.1.0"
description = "Multi-Agent Support System"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.50",
    "langchain>=0.3.10",
    "langchain-openai>=0.2.8",
    "langchain-elasticsearch>=0.3.0",
    "langchain-google-community>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "elasticsearch>=8.17.0",
    "psycopg2-binary>=2.9.10",
    "click>=8.1.8",
    "rich>=13.9.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "websocket-client>=1.8.0",
    "requests>=2.32.0",
    "urllib3>=2.2.0",
    "docling>=2.40.0",
    "ipykernel>=6.29.5",
    "torch>=2.7.1",
    "transformers>=4.53.1",
    "docling-core>=2.40.0",
    "pymupdf>=1.26.3",
    "pillow>=11.3.0",
    "huggingface-hub>=0.33.2",
    "langgraph-checkpoint-postgres>=2.0.21",
    "psycopg-binary>=3.2.9",
    "langchain-postgres>=0.0.15",
    "langchain-core>=0.3.68",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
support = "src.support_system.run:cli"
